Fix issue where ssh_keyfile could cause trace

+++ hot/shell.py
@@ -241,7 +241,8 @@ def run_resource_tests(hc, stack_id, resource_tests):
             hot.tests.script.run_script(test_name, test[test_name])
         else:
             print "  No tests defined."
-    hot.utils.files.delete_file(ssh_key_file)
+    if 'ssh_key_file' in resource_tests:
+        hot.utils.files.delete_file(ssh_key_file)
 
 
 def update_dict(items, outputs):
